chore: empty default implementation for RenderSystem interface
Requires https://github.com/MovingBlocks/Terasology/pull/4789
